test: move tests directory to root
